Added seller's rating Fixed ad.html (except buttons) Need to add chat

+++ src/main/java/com/netcracker/auto/entity/User.java
@@ -34,9 +34,9 @@ public class User {
     @Column(name = "balance")
     private double balance;
     @Column(name = "seller_rating")
-    private int seller_rating;
+    private Double seller_rating;
     @Column(name = "buyers_rating")
-    private int buyers_rating;
+    private Double buyers_rating;
     @Column(name = "password")
     private String password;
     @Column(name = "images", length = 64)
@@ -52,12 +52,12 @@ public class User {
         return getRoles();
     }
 
-    public void setDefault(){
+    public void setDefault() {
         confirmed = false;
-        date_of_birth =  LocalDate.now();
+        date_of_birth = LocalDate.now();
         balance = 0;
-        buyers_rating = 0;
-        seller_rating = 0;
+        buyers_rating = 0.0;
+        seller_rating = 0.0;
         phone_number = "";
     }
 
@@ -66,11 +66,12 @@ public class User {
         if (images == null || userId == null) return null;
         return "/user-photos/" + userId + "/" + images;
     }
-    public String getFullName(){
-        return last_name+" "+first_name;
+
+    public String getFullName() {
+        return last_name + " " + first_name;
     }
 
-    public void addRole(RolesEntity rolesEntity){
+    public void addRole(RolesEntity rolesEntity) {
         this.roles.add(rolesEntity);
     }
 
@@ -118,7 +119,7 @@ public class User {
         return balance;
     }
 
-    public int getBuyers_rating() {
+    public Double getBuyers_rating() {
         return buyers_rating;
     }
 
@@ -126,7 +127,7 @@ public class User {
         return userId;
     }
 
-    public int getSeller_rating() {
+    public Double getSeller_rating() {
         return seller_rating;
     }
 
@@ -146,7 +147,7 @@ public class User {
         this.balance = balance;
     }
 
-    public void setBuyers_rating(int buyers_rating) {
+    public void setBuyers_rating(Double buyers_rating) {
         this.buyers_rating = buyers_rating;
     }
 
@@ -174,7 +175,7 @@ public class User {
         this.userId = userId;
     }
 
-    public void setSeller_rating(int seller_rating) {
+    public void setSeller_rating(Double seller_rating) {
         this.seller_rating = seller_rating;
     }
